requirements.txt
Flask
couchbase
python-dotenv
pytest